// server/@types/shared/models/PlacementRequestSortField.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PlacementRequestSortField = 'duration' | 'expected_arrival' | 'created_at' | 'application_date' | 'request_type' | 'person_name' | 'person_risks_tier' | 'name' | 'canonical_arrival_date';